[220201-LHG] 패키지 구조 수정 - domain, web 분리
