+++ src/main/java/com/ootd/with/domain/hashtag/PostHashtagId.java
@@ -1,4 +1,4 @@
-package com.ootd.with.domain;
+package com.ootd.with.domain.hashtag;
 
 import lombok.EqualsAndHashCode;
 import lombok.NoArgsConstructor;
