+++ src/main/java/com/ootd/with/domain/likey/CommentLikey.java
@@ -1,4 +1,6 @@
-package com.ootd.with.domain;
+package com.ootd.with.domain.likey;
+
+import com.ootd.with.domain.comment.Comment;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -6,10 +8,9 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 @Entity
-public class PostLikey extends Likey {
+public class CommentLikey extends Likey {
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "post_id")
-    private Post post;
+    @JoinColumn(name = "comment_id")
+    private Comment comment;
 }
-
