+++ src/main/java/com/ootd/with/domain/likey/PostLikey.java
@@ -1,4 +1,6 @@
-package com.ootd.with.domain;
+package com.ootd.with.domain.likey;
+
+import com.ootd.with.domain.post.Post;
 
 import javax.persistence.Entity;
 import javax.persistence.FetchType;
@@ -6,9 +8,10 @@ import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
 
 @Entity
-public class CommentLikey extends Likey {
+public class PostLikey extends Likey {
 
     @ManyToOne(fetch = FetchType.LAZY)
-    @JoinColumn(name = "comment_id")
-    private Comment comment;
+    @JoinColumn(name = "post_id")
+    private Post post;
 }
+
